POLIO-1556: move Evaluation files out of rounds folder
